requirements.txt
aiohttp==3.7.3
python-dotenv==0.15.0
Telethon==1.19.2
youtube-dl
